requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
loguru==0.6.0
pydantic==1.10.2
PyMySQL==1.0.2
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
win32-setctime==1.1.0
